Program is tested on serial communications. Setting train functions and speed appear to be working. The Fifo buffer is also working as intended
The program will scream out over serial port when the Fifo buffer
overflows.

Many serial texts are yet to be removed, the DCC signal is not tested.

The next logical step is to build in a step functionality, in which 1
DCC packet will be clocked out at the time. Than we can verify the
packet's integrities. The flow of packets must also be observed, though
this should be fine.

The double array should also be debugged before I test with actual
trains. I do have faith that they work.